pyproject.toml
[project]
name = "recurse-ask-zulip"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
    "zulip>=0.9.1",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
